01 authentication

+++ source/server/pkg/users/handler.go
@@ -6,12 +6,12 @@ import (
 	"log"
 	"net/http"
 
+	"github.com/PfMartin/weGoNice/server/pkg/auth"
+	"github.com/PfMartin/weGoNice/server/pkg/models"
 	"github.com/gorilla/mux"
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/bson/primitive"
 	"go.mongodb.org/mongo-driver/mongo"
-
-	"github.com/PfMartin/weGoNice/server/pkg/auth"
 )
 
 type Handler struct {
@@ -25,24 +25,29 @@ func NewHandler(db *mongo.Client) Handler {
 }
 
 func (h *Handler) GetAllUsers(w http.ResponseWriter, r *http.Request) {
+	if !auth.IsAdminContextOk(r) {
+		http.Error(w, "Not authorized to view all users", http.StatusUnauthorized)
+		return
+	}
 	coll := h.DB.Database(h.dbName).Collection(h.collection)
 
 	cursor, err := coll.Find(context.TODO(), bson.D{})
-
 	if err != nil {
-		log.Printf("Error: Couldn't read data: %v", err)
+		log.Printf("Error: Failed to find users: %v", err)
+		http.Error(w, "Failed to find users", http.StatusNotFound)
+		return
 	}
 
-	var users []User
+	var users []models.User
 	if err = cursor.All(context.TODO(), &users); err != nil {
-		log.Printf("Error: Couldn't parse users, %v", err)
+		log.Printf("Error: Failed to parse users, %v", err)
+		http.Error(w, "Failed to parse users", http.StatusInternalServerError)
+		return
 	}
 
 	w.Header().Add("Content-Type", "application/json")
 	w.WriteHeader(http.StatusOK)
 	json.NewEncoder(w).Encode(users)
-
-	logSuccess("getAll", "none")
 }
 
 func (h *Handler) GetUserById(w http.ResponseWriter, r *http.Request) {
@@ -50,58 +55,90 @@ func (h *Handler) GetUserById(w http.ResponseWriter, r *http.Request) {
 
 	objectId, err := primitive.ObjectIDFromHex(id)
 	if err != nil {
-		log.Printf("Error: Could not parse id to objectId: %v", err)
+		log.Printf("Error: Failed to parse id to ObjectID: %v", err)
+		http.Error(w, "Failed to parse id to ObjectID", http.StatusBadRequest)
+		return
 	}
 
 	filter := bson.M{"_id": objectId}
 
 	coll := h.DB.Database(h.dbName).Collection(h.collection)
 
-	var user User
+	var user models.User
 	err = coll.FindOne(context.TODO(), filter).Decode(&user)
 	if err != nil {
-		log.Printf("Error: Couldn't find a user")
+		log.Printf("Error: Failed to find a user")
+		http.Error(w, "Failed to find user", http.StatusNotFound)
+		return
+	}
+
+	if !auth.IsEmailContextOk(user.Email, r) {
+		http.Error(w, "Not authorized to see details about another user", http.StatusUnauthorized)
+		return
 	}
 
 	w.Header().Add("Content-Type", "application/json")
 	w.WriteHeader(http.StatusOK)
 	json.NewEncoder(w).Encode(user)
-
-	logSuccess("getById", objectId.String())
 }
 
 func (h *Handler) CreateUser(w http.ResponseWriter, r *http.Request) {
-	var payload User
-
-	decoder := json.NewDecoder(r.Body)
-	decoder.DisallowUnknownFields()
-	err := decoder.Decode(&payload)
-
-	if err != nil {
-		log.Printf("Could not decode body: %v", err)
-	}
-
-	hashedPassword, err := auth.HashPassword(payload.Password)
-
-	if err != nil {
-		log.Printf("Failed to hash password: %s", err)
+	if !auth.IsAdminContextOk(r) {
+		http.Error(w, "Not authorized to create user", http.StatusUnauthorized)
 		return
 	}
 
-	data := bson.D{{Key: "lastname", Value: payload.Lastname}, {Key: "firstname", Value: payload.Firstname}, {Key: "email", Value: payload.Email}, {Key: "password", Value: hashedPassword}}
-
-	coll := h.DB.Database(h.dbName).Collection(h.collection)
-	result, err := coll.InsertOne(context.TODO(), data)
+	var user models.User
+	decoder := json.NewDecoder(r.Body)
+	decoder.DisallowUnknownFields()
+	err := decoder.Decode(&user)
 
 	if err != nil {
-		log.Printf("Error: Couldn't insert data: %v", err)
+		log.Printf("Error: Failed to decode body: %v", err)
+		http.Error(w, "Failed to decode body", http.StatusBadRequest)
+		return
 	}
+
+	coll := h.DB.Database(h.dbName).Collection(h.collection)
+
+	// Check if user already exists
+	var existingUser models.User
+	filter := bson.D{{Key: "email", Value: user.Email}}
+	err = coll.FindOne(context.TODO(), filter).Decode(&existingUser)
+	if err == nil {
+		log.Printf("Error: User with email %s already exists: %v", user.Email, err)
+		w.Header().Add("Content-Type", "application/json")
+		w.WriteHeader(http.StatusNotAcceptable)
+		msg := "User with email " + user.Email + " already exists."
+
+		response := map[string]string{
+			"id": msg,
+		}
+		json.NewEncoder(w).Encode(response)
+		return
+	}
+
+	// Hash password and create new user
+	hashedPassword, err := auth.HashPassword(user.Password)
+	if err != nil {
+		log.Printf("Error: Failed to hash password: %s", err)
+		http.Error(w, "Failed to hash password", http.StatusInternalServerError)
+		return
+	}
+
+	data := bson.D{{Key: "lastname", Value: user.Lastname}, {Key: "firstname", Value: user.Firstname}, {Key: "email", Value: user.Email}, {Key: "password", Value: hashedPassword}}
+	cursor, err := coll.InsertOne(context.TODO(), data)
+	if err != nil {
+		log.Printf("Error: Failed insert data: %v", err)
+		http.Error(w, "Failed to insert data", http.StatusInternalServerError)
+		return
+	}
+
+	userId := cursor.InsertedID.(primitive.ObjectID)
 
 	w.Header().Add("Content-Type", "application/json")
 	w.WriteHeader(http.StatusCreated)
-	json.NewEncoder(w).Encode(result)
-
-	logSuccess("add", "none")
+	json.NewEncoder(w).Encode(userId)
 }
 
 func (h *Handler) UpdateUserById(w http.ResponseWriter, r *http.Request) {
@@ -109,48 +146,52 @@ func (h *Handler) UpdateUserById(w http.ResponseWriter, r *http.Request) {
 
 	objectId, err := primitive.ObjectIDFromHex(id)
 	if err != nil {
-		log.Printf("Error: Could not parse id to objectId: %v", err)
+		log.Printf("Error: Failed to parse id to ObjectID: %v", err)
+		http.Error(w, "Failed to parse id to ObjectID", http.StatusInternalServerError)
+		return
 	}
 
-	var user User
+	var user models.User
 
 	decoder := json.NewDecoder(r.Body)
 	decoder.DisallowUnknownFields()
 	err = decoder.Decode(&user)
-
 	if err != nil {
-		log.Printf("Could not decode body: %v", err)
+		log.Printf("Error: Failed to decode user: %v", r.Body)
+		http.Error(w, "Failed to decode user", http.StatusBadRequest)
+		return
+	}
+
+	if !auth.IsEmailContextOk(user.Email, r) {
+		http.Error(w, "Not authorized to update other user", http.StatusUnauthorized)
+		return
 	}
 
 	filter := bson.M{"_id": objectId}
-	// Email and Password should be updated by hitting dedicated endpoints because these are the login credentials
 	update := bson.M{"$set": bson.M{
-		"lastname": user.Lastname,
+		"lastname":  user.Lastname,
 		"firstname": user.Firstname,
 	}}
-
 	coll := h.DB.Database(h.dbName).Collection(h.collection)
 	result, err := coll.UpdateOne(context.TODO(), filter, update)
 	if err != nil {
-		log.Printf("Error: Couldn't update user: %v", err)
+		log.Printf("Error: Failed to update user: %v", err)
+		http.Error(w, "Failed to update user", http.StatusInternalServerError)
 		return
 	}
 
 	w.Header().Add("Content-Type", "application/json")
 	w.WriteHeader(http.StatusOK)
 	json.NewEncoder(w).Encode(result)
-
-	logSuccess("updateById", objectId.String())
 }
 
 func (h *Handler) DeleteUserById(w http.ResponseWriter, r *http.Request) {
 	id := mux.Vars(r)["id"]
 
 	objectId, err := primitive.ObjectIDFromHex(id)
-
 	if err != nil {
-		log.Printf("Error: Could not parse id to objectId: %v", err)
-		w.WriteHeader(http.StatusBadRequest)
+		log.Printf("Error: Failed to parse id to objectId: %v", err)
+		http.Error(w, "Failed to parse id to ObjectID", http.StatusBadRequest)
 		return
 	}
 
@@ -158,47 +199,44 @@ func (h *Handler) DeleteUserById(w http.ResponseWriter, r *http.Request) {
 
 	coll := h.DB.Database(h.dbName).Collection(h.collection)
 
+	var user models.User
+	err = coll.FindOne(context.TODO(), filter).Decode(&user)
+	if err != nil {
+		log.Printf("Error: Failed to find user: %v", err)
+		http.Error(w, "Failed to find user", http.StatusNotFound)
+		return
+	}
+
+	if !auth.IsEmailContextOk(user.Email, r) {
+		http.Error(w, "Not authorized to delete other user", http.StatusUnauthorized)
+		return
+	}
+
+	filter = bson.M{"_id": objectId}
+	coll = h.DB.Database(h.dbName).Collection(h.collection)
 	coll.DeleteOne(context.TODO(), filter)
 
 	w.Header().Add("Content-Type", "application/json")
 	w.WriteHeader(http.StatusOK)
 	json.NewEncoder(w).Encode("Deleted user")
-
-	logSuccess("deleteById", objectId.String())
 }
 
 func (h *Handler) DeleteAllUsers(w http.ResponseWriter, r *http.Request) {
+	if !auth.IsAdminContextOk(r) {
+		http.Error(w, "Not authorized to delete all users", http.StatusUnauthorized)
+		return
+	}
+
 	coll := h.DB.Database(h.dbName).Collection(h.collection)
 
 	result, err := coll.DeleteMany(context.TODO(), bson.D{})
 	if err != nil {
-		log.Printf("Error: Couldn't find a user")
+		log.Printf("Error: Failed to delete all users: %v", err)
+		http.Error(w, "Failed to delete all users", http.StatusInternalServerError)
+		return
 	}
 
 	w.Header().Add("Content-Type", "application/json")
 	w.WriteHeader(http.StatusOK)
 	json.NewEncoder(w).Encode(result.DeletedCount)
-
-	logSuccess("deleteAll", "none")
-}
-
-func logSuccess(operation string, objectId string) {
-	var operationText string
-
-	switch operation {
-	case "getAll":
-		operationText = "returned all users"
-	case "getById":
-		operationText = "returned user with id: " + objectId
-	case "add":
-		operationText = "added user"
-	case "deleteAll":
-		operationText = "deleted all users"
-	case "deleteById":
-		operationText = "deleted user with id: " + objectId
-	case "updateById":
-		operationText = "updated user with id: " + objectId
-	}
-
-	log.Printf("Successfully %s\n", operationText)
 }
